Decouple service processing from components/entry/EntryForm.tsx
