+++ frontend/src/components/entry/EntryForm.tsx
@@ -23,12 +23,9 @@ import { EditableEntry, EditableEntryAttrs } from "./entryForm/EditableEntry";
 
 import { EditAttributeValue } from "components/entry/entryForm/EditAttributeValue";
 import { DjangoContext } from "services/DjangoContext";
-
-interface asArrayNamedObjectBoolean {
-  [key: string]: EntryAttributeValueObject & {
-    boolean: boolean;
-  };
-}
+import {
+  updateEntryInfoValueFromValueInfo,
+} from "services/entry/Edit";
 
 interface Props {
   entryInfo: EditableEntry;
@@ -63,145 +60,10 @@ export const EntryForm: FC<Props> = ({
     attrType: number,
     valueInfo: any
   ) => {
-    switch (attrType) {
-      case djangoContext?.attrTypeValue.date:
-      case djangoContext?.attrTypeValue.string:
-      case djangoContext?.attrTypeValue.text:
-        entryInfo.attrs[attrName].value.asString = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
+    updateEntryInfoValueFromValueInfo(entryInfo, attrName, attrType, valueInfo);
 
-      case djangoContext?.attrTypeValue.boolean:
-        entryInfo.attrs[attrName].value.asBoolean = valueInfo.checked;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.object:
-        entryInfo.attrs[attrName].value.asObject = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.group:
-        entryInfo.attrs[attrName].value.asGroup = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.role:
-        entryInfo.attrs[attrName].value.asRole = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.named_object:
-        if (valueInfo?.key !== undefined) {
-          entryInfo.attrs[attrName].value.asNamedObject = {
-            [valueInfo.key]:
-              Object.values(
-                entryInfo.attrs[attrName].value.asNamedObject ?? {}
-              )[0] ?? null,
-          };
-        } else {
-          entryInfo.attrs[attrName].value.asNamedObject = {
-            [Object.keys(
-              entryInfo.attrs[attrName].value.asNamedObject ?? {}
-            )[0] ?? ""]: valueInfo.value,
-          };
-        }
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_string:
-        if (entryInfo.attrs[attrName].value?.asArrayString == null) {
-          entryInfo.attrs[attrName].value.asArrayString = [];
-        }
-        // @ts-ignore
-        entryInfo.attrs[attrName].value.asArrayString[valueInfo.index] =
-          valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_object:
-        entryInfo.attrs[attrName].value.asArrayObject = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_group:
-        entryInfo.attrs[attrName].value.asArrayGroup = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_role:
-        entryInfo.attrs[attrName].value.asArrayRole = valueInfo.value;
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_named_object:
-        if (
-          entryInfo.attrs[attrName].value.asArrayNamedObject?.length ??
-          0 <= valueInfo.index
-        ) {
-          entryInfo.attrs[attrName].value.asArrayNamedObject?.push({
-            // @ts-ignore
-            "": null,
-          });
-          changeAttributes({ ...entryInfo.attrs });
-          break;
-        }
-
-        if (valueInfo?.key !== undefined) {
-          // @ts-ignore
-          entryInfo.attrs[attrName].value.asArrayNamedObject[valueInfo.index] =
-            {
-              [valueInfo.key]:
-                Object.values(
-                  entryInfo.attrs[attrName].value.asArrayNamedObject?.[
-                    valueInfo.index
-                  ] ?? {}
-                )[0] ?? null,
-            };
-        } else {
-          // @ts-ignore
-          entryInfo.attrs[attrName].value.asArrayNamedObject[valueInfo.index] =
-            {
-              [Object.keys(
-                entryInfo.attrs[attrName].value.asArrayNamedObject?.[
-                  valueInfo.index
-                ] ?? {}
-              )[0] ?? ""]: {
-                ...Object.values(
-                  entryInfo.attrs[attrName].value.asArrayNamedObject?.[
-                    valueInfo.index
-                  ] ?? {}
-                )[0],
-                id: valueInfo.value.id,
-                name: valueInfo.value.name,
-              },
-            };
-        }
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-
-      case djangoContext?.attrTypeValue.array_named_object_boolean:
-        (
-          entryInfo.attrs[attrName].value
-            .asArrayNamedObject as asArrayNamedObjectBoolean[]
-        )[valueInfo.index] = {
-          [Object.keys(
-            entryInfo.attrs[attrName].value.asArrayNamedObject?.[
-              valueInfo.index
-            ] ?? {}
-          )[0] ?? ""]: {
-            ...Object.values(
-              entryInfo.attrs[attrName].value.asArrayNamedObject?.[
-                valueInfo.index
-              ] ?? {}
-            )[0],
-            boolean: valueInfo.checked,
-          },
-        };
-
-        changeAttributes({ ...entryInfo.attrs });
-        break;
-    }
+    // Update entryInfo.attrs value depends on the changing values
+    changeAttributes({ ...entryInfo.attrs });
   };
 
   const handleClickAddListItem = (
